requirements.txt
speedtest-cli
schedule
matplotlib
pandas